+++ ARKBreedingStats/species/CreatureColors.cs
@@ -12,18 +12,18 @@ namespace ARKBreedingStats.species
         /// </summary>
         /// <param name="colorId"></param>
         /// <returns></returns>
-        public static string CreatureColorName(int colorId) => values.Values.V.Colors.ById(colorId).Name;
+        public static string CreatureColorName(byte colorId) => values.Values.V.Colors.ById(colorId).Name;
 
         /// <summary>
         /// Returns the Color struct of an ArkColor by id.
         /// </summary>
         /// <param name="colorId"></param>
         /// <returns></returns>
-        public static Color CreatureColor(int colorId) => values.Values.V.Colors.ById(colorId).Color;
+        public static Color CreatureColor(byte colorId) => values.Values.V.Colors.ById(colorId).Color;
 
         /// <summary>
         /// Returns the ArkColor by id.
         /// </summary>
-        public static ArkColor CreatureArkColor(int colorId) => values.Values.V.Colors.ById(colorId);
+        public static ArkColor CreatureArkColor(byte colorId) => values.Values.V.Colors.ById(colorId);
     }
 }
